Refactored SegregatedListAllocator out
My thoughts are that it is more important right now to get the project into something I can submit - a finished state. I will do all the other things I need to do (which will (hopefully not but) take longer than I think) and then come back to this. If not, it can always just be something I write about rather than actually implement.

The reason it was slow is that it was getting stuck in FreeListNode::AddFreeListNode's while loop. This is because we were just adding loads of free list nodes. If we come back to this allocator, then we will need the coalescing logic as well.

Anyway, this was worth looking into as we can now write about this on our dissertation more knowledgably. However as I say I don't think its worth doing now. We will finish up the other stuff we have to do (and honestly I think what we have done already is more than enough...) and come back to it if we have time. We are now starting to get set other assignments as well, so our attention will start to be divided.

Also - remember! We are going for an *easy* and *general purpose* allocator! Easy for the user to use. The Free List Allocator is that